requirements.txt
streamlit
transformers
diffusers
torch
torchvision
Pillow
accelerate
